refactor(cli): split doctor_cmd into a package
The doctor command was a single ~995-line module: a giant single-repo
check battery, a workspace check battery, advisory helpers, and the
Click entrypoint all in one file, with the repo-check function alone
running ~460 lines and six nested async closures.

Split it into a doctor_cmd/ package:

- _types.py           DoctorCheck row type and small formatting helpers
- advisories.py       advisory-only signals (never flip pass/fail)
- repo_checks.py      the single-repo check battery + distill checks
- workspace_checks.py workspace validation + MCP registration scan
- command.py          the Click entrypoint and dispatch

Pure structural move: symbols relocate verbatim, lazy imports stay
function-local, and __init__.py re-exports the public surface so main.py
and every test importer are unchanged. The console monkeypatch targets
in test_doctor_cli_version.py and test_freshness_stamp_fixes.py are
repointed to doctor_cmd.advisories to match the advisory helpers' new
home (patch address only, no test logic changed).
